Fixes in some use cases and optimizations
