+++ app/src/main/java/com/weatherxm/data/Modules.kt
@@ -165,6 +165,7 @@ import com.weatherxm.ui.claimdevice.location.ClaimLocationViewModel
 import com.weatherxm.ui.claimdevice.pulse.ClaimPulseViewModel
 import com.weatherxm.ui.claimdevice.wifi.ClaimWifiViewModel
 import com.weatherxm.ui.connectwallet.ConnectWalletViewModel
+import com.weatherxm.ui.deeplinkrouteractivity.DeepLinkRouterViewModel
 import com.weatherxm.ui.deleteaccount.DeleteAccountViewModel
 import com.weatherxm.ui.deleteaccountsurvey.DeleteAccountSurveyViewModel
 import com.weatherxm.ui.devicedetails.DeviceDetailsViewModel
@@ -198,7 +199,6 @@ import com.weatherxm.ui.rewardslist.RewardsListViewModel
 import com.weatherxm.ui.signup.SignupViewModel
 import com.weatherxm.ui.startup.StartupViewModel
 import com.weatherxm.ui.updateprompt.UpdatePromptViewModel
-import com.weatherxm.ui.deeplinkrouteractivity.DeepLinkRouterViewModel
 import com.weatherxm.ui.widgets.selectstation.SelectStationViewModel
 import com.weatherxm.usecases.AnalyticsOptInUseCase
 import com.weatherxm.usecases.AnalyticsOptInUseCaseImpl
@@ -214,10 +214,6 @@ import com.weatherxm.usecases.ChartsUseCase
 import com.weatherxm.usecases.ChartsUseCaseImpl
 import com.weatherxm.usecases.ClaimDeviceUseCase
 import com.weatherxm.usecases.ClaimDeviceUseCaseImpl
-import com.weatherxm.usecases.WalletUseCase
-import com.weatherxm.usecases.WalletUseCaseImpl
-import com.weatherxm.usecases.DeleteAccountSurveyUseCase
-import com.weatherxm.usecases.DeleteAccountSurveyUseCaseImpl
 import com.weatherxm.usecases.DeleteAccountUseCase
 import com.weatherxm.usecases.DeleteAccountUseCaseImpl
 import com.weatherxm.usecases.DeviceDetailsUseCase
@@ -234,8 +230,6 @@ import com.weatherxm.usecases.ForecastUseCase
 import com.weatherxm.usecases.ForecastUseCaseImpl
 import com.weatherxm.usecases.HistoryUseCase
 import com.weatherxm.usecases.HistoryUseCaseImpl
-import com.weatherxm.usecases.PasswordPromptUseCase
-import com.weatherxm.usecases.PasswordPromptUseCaseImpl
 import com.weatherxm.usecases.PreferencesUseCase
 import com.weatherxm.usecases.PreferencesUseCaseImpl
 import com.weatherxm.usecases.RemoteBannersUseCase
@@ -250,6 +244,8 @@ import com.weatherxm.usecases.StatsUseCase
 import com.weatherxm.usecases.StatsUseCaseImpl
 import com.weatherxm.usecases.UserUseCase
 import com.weatherxm.usecases.UserUseCaseImpl
+import com.weatherxm.usecases.WalletUseCase
+import com.weatherxm.usecases.WalletUseCaseImpl
 import com.weatherxm.usecases.WidgetCurrentWeatherUseCase
 import com.weatherxm.usecases.WidgetCurrentWeatherUseCaseImpl
 import com.weatherxm.usecases.WidgetSelectStationUseCase
@@ -537,14 +533,8 @@ private val usecases = module {
     single<PreferencesUseCase> {
         PreferencesUseCaseImpl(get(), get(), get())
     }
-    single<DeleteAccountSurveyUseCase> {
-        DeleteAccountSurveyUseCaseImpl(get())
-    }
     single<DeleteAccountUseCase> {
         DeleteAccountUseCaseImpl(get(), get())
-    }
-    single<PasswordPromptUseCase> {
-        PasswordPromptUseCaseImpl(get())
     }
     single<BluetoothScannerUseCase> {
         BluetoothScannerUseCaseImpl(get())
