+++ app/src/main/java/com/weatherxm/ui/passwordprompt/PasswordPromptViewModel.kt
@@ -3,17 +3,17 @@ package com.weatherxm.ui.passwordprompt
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.weatherxm.R
+import com.weatherxm.analytics.AnalyticsWrapper
 import com.weatherxm.data.Resource
 import com.weatherxm.data.SingleLiveEvent
-import com.weatherxm.usecases.PasswordPromptUseCase
-import com.weatherxm.analytics.AnalyticsWrapper
+import com.weatherxm.usecases.AuthUseCase
 import com.weatherxm.util.Failure.getDefaultMessage
 import com.weatherxm.util.Resources
 import com.weatherxm.util.Validator
 import kotlinx.coroutines.launch
 
 class PasswordPromptViewModel(
-    private val usecase: PasswordPromptUseCase,
+    private val usecase: AuthUseCase,
     private val resources: Resources,
     private val analytics: AnalyticsWrapper
 ) : ViewModel() {
